Move generator loading functions into a concern
